+++ inst/templates/ad_advs.R
@@ -20,12 +20,6 @@ data("adsl")
 
 vs <- convert_blanks_to_na(vs)
 
-# The CDISC Pilot Data contains no SUPPVS data
-# If you have a SUPPVS then uncomment function below
-
-# vs <- derive_vars_suppqual(vs, suppvs) %>%
-
-
 # ---- Lookup tables ----
 
 # Assign PARAMCD, PARAM, and PARAMN